Wrap collector in a service
Updated the "collector" process that gets the network information and
then stores it in the database to run as service process.
Refactored how the Poller queries the network so that I can initialize
the `Poller` once from a config object and use that with different OLSR
data each time in the collector service.
Updated some functions/classes to take specific arguments rather than
configuration objects.