Use Gunicorn WSGI server
Waitress has had issues of hanging, with over a dozen waiting threads.
Switch to Gunicorn (with it's processor workers) so that we can directly
call `rrdtool.graphv()` without need to wrap it in a process.
Use Pyramid's `prepare()` instead of `bootstrap()` for the CLI process
since we are not using PasteDeploy any longer to configure the WSGI app.