Save node errors during network collection to database
Add basic reverse DNS lookup functionality and use it to lookup names of
nodes that having polling failures.
Refactor poller functionality from `network_info` funcion into `Poller`
class because the former required passing various settings through
multiple levels of the stack.
Wrap node responses in a common `NodeResult` data class, which (almost)
always has the node's name as well as the IP address, both for errors
and successes.
This eliminated the need for a "name to IP" cache and meant the "IP to
name" lookup could be built where it was needed.  Thus, more of the
intermediate steps for polling now return a list of results  instead of
named tuples.
This required updating both the `collector` and the `report` modules.
Store node IP/name and response to database.
Modernize type hints.