Render graphs using "client" timezone
Update `request.timezone` to be timezone object (so Pendulum doesn't
need to convert string back to object when rendering).
Wrap RRDtool graph calls in changes to the `TZ` environment variable so
they render in the preferred timezone.
Update graph views to pass timezone to the graph rendering functions.
Fix bug in the fallback for loading timezone from cookie.