Refactor topology collection
Simplify building the topology information from OLSR data by building a
model of the data instead of creating asynchronous generators.
Build the link data during the initial looping through the data, instead
of looping through a second time.
Generalized the OLSR data as topology, because there is potentially an
API for getting this data as well.